refactor: Move Numba kernels to a dedicated module
